sunxi: alter dryrun in build-image.sh
Adopt a standard of the dryrun variable either being empty or non-empty,
with non-empty meaning dryrun is active. This allows the invocation of
other commands with dryrun mode using ${dryrun:+--flag} where --flag is
whatever that program's dryrun flag is.

e.g. mktemp takes -u as dryrun, and even in dryrun, you still want to
run mktemp. This method of managing the variable makes it easier.