sunxi: Find rootfs, dtb and kernel in parent chain
Stop requiring that the rootfs image, dtb and kernel are in the ultimate
board directory and search for them through the chain. For this to work,
the ultimate build-image script needs to call the sunxi build-image
script with the rootfs layer name and the device tree name.

build-image.sh will locate the images and symlink them into the image
dir ready for genimage to find.