sunxi: Build linux kernel once for sunxi only
Move the build of the linux kernel for sunxi kernels to the 'sunxi'
board. We built the kernel previously per-board as each board had a
different DTB from the kernel. However the DTB config accepts a list of
DTBs, so list all our sunxi board dtbs there and move the build to the
sunxi board.

The kernel across the sunxi boards was always the same, so this makes no
difference. As build-image.sh can now locate the dtb and not require it
to be in the ultimate board image dir, this can be done.