sunxi: Move final image build to `sunxi` board
The final image can be built in the sunxi board instead of each of the
ultimate boards as the filesystem is the same across all the sunxi
boards. The final image is just the rootfs. The SD card et al images
still need to be build by the boards as they contain board-specific
uboot code and DTB. As the rootfs contains nothing from uboot, the
rootfs can be built without knowledge of any specific boards.

Split final into a sunxi final and a board-specific genimage layer.
final builds the rootfs and genimage builds the sdcard images.